build(docker): remove symbolic links for node and npm binaries
This reverts the addition of symbolic links for node and npm binaries in the Dockerfile, as they are no longer needed after switching to direct usage of nodejs and npm paths. The change reduces filesystem modifications and simplifies the container setup process.

See commit 57cf4fe for the original change being reverted.